+++ lib/json_matchers/matcher.rb
@@ -8,11 +8,9 @@ module JsonMatchers
     end
 
     def matches?(response)
-      @response = response
-
       JSON::Validator.validate!(
         schema_path.to_s,
-        response.body,
+        json_from(response).to_s,
         options,
       )
     rescue JSON::Schema::ValidationError => ex
@@ -29,5 +27,13 @@ module JsonMatchers
     private
 
     attr_reader :schema_path, :options
+
+    def json_from(response)
+      if response.respond_to?(:body)
+        response.body
+      else
+        response
+      end
+    end
   end
 end