+++ spec/json_matchers/match_response_schema_spec.rb
@@ -39,6 +39,22 @@ describe JsonMatchers, "#match_response_schema" do
       to match_response_schema("foo_schema", strict: false)
   end
 
+  it "validates a JSON string" do
+    create_schema("foo_schema", {
+      "type" => "object",
+      "required" => [
+        "id",
+      ],
+      "properties" => {
+        "id" => { "type" => "number" },
+      },
+      "additionalProperties" => false,
+    })
+
+    expect(response_for({ "id" => 1 }).body).
+      to match_response_schema("foo_schema")
+  end
+
   it "fails when the body contains a property with the wrong type" do
     create_schema("foo_schema", {
       "type" => "object",