+++ NEWS.md
@@ -1,6 +1,10 @@
 master
 ======
 
+* Accept a JSON string along with a `response`.
+
+[#43]: https://github.com/thoughtbot/json_matchers/pull/43
+
 0.5.1
 =====